\name{DAVIDQuery-package}
\alias{DAVIDQuery-package}
\docType{package}
\title{
Retrieval from DAVID bioinformatics data resource.
}
\description{
Tools to retrieve data from  DAVID, the Database for Annotation, Visualization and Integrated Discovery.
}
\details{
\tabular{ll}{
Package: \tab DAVIDQuery\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2008-10-31\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
Vignette and overview: \tab vignette("DAVIDQuery")\cr
}
}
\author{	
Roger Day <day@upci.pitt.edu>
}
\references{
Home base for DAVID  (in this package as DAVIDURLBase):
\url{http://david.abcc.ncifcrf.gov}

Huang et al, DAVID gene ID conversion:
\url{ http://www.pubmedcentral.nih.gov/articlerender.fcgi?tool=pubmed&pubmedid=18841237}}
\keyword{ package }
\keyword{ database }

