\name{removeHTMLComments}
\alias{removeHTMLComments}
\title{Remove HTML comments from character string representing the HTML page...}
\usage{removeHTMLComments(string)}
\description{Remove HTML comments from character string representing the HTML page}
\value{character string with HTML comments ('<!--...-->') removed}
\keyword{internal}
\author{Roger Day, Alex Lisovich}
\arguments{\item{string}{character string representing HTML page}}
