\name{DAVIDToolChoices}
\alias{DAVIDToolChoices}
\alias{DAVIDTypeChoices}
\alias{DAVIDAnnotChoices}
\docType{data}
\title{ Choices for the DAVID query  parameters }
\description{
	DAVIDToolChoices, DAVIDTypeChoices,  and DAVIDAnnotChoices are string vectors used to construct pick menus, when the corresponding arguments to DAVIDQuery are not provided.
}
\details{
	The source of these lists can be found at
	\url{http://david.abcc.ncifcrf.gov/content.jsp?file=DAVID_API.html#input_list}
	The lists posted there may change in the future.
	No mechanism is included in this package to retrieve and parse the page.
}
\source{
	\url{http://david.abcc.ncifcrf.gov/content.jsp?file=DAVID_API.html#input_list}
}
\seealso{ \code{\link{DAVIDQuery}} }
\keyword{ database }
